+++ app/src/main/java/com/axel_stein/glucose_tracker/ui/statistics/StatisticsFragment.kt
@@ -9,15 +9,14 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.axel_stein.glucose_tracker.R
-import com.axel_stein.glucose_tracker.utils.hideView
-import com.axel_stein.glucose_tracker.utils.setItemSelectedListener
-import com.axel_stein.glucose_tracker.utils.show
+import com.axel_stein.glucose_tracker.utils.*
 
 class StatisticsFragment: Fragment() {
-    private val model: StatisticsViewModel by viewModels()
+    private val viewModel: StatisticsViewModel by viewModels()
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         val root = inflater.inflate(R.layout.fragment_statistics, container, false)
+        val content = root.findViewById<View>(R.id.content)
         val spinnerPeriod = root.findViewById<Spinner>(R.id.spinner_period)
         val textMin = root.findViewById<TextView>(R.id.text_min)
         val textMax = root.findViewById<TextView>(R.id.text_max)
@@ -26,14 +25,28 @@ class StatisticsFragment: Fragment() {
         val textControlGood = root.findViewById<TextView>(R.id.text_a1c_control_good)
         val textControlAvg = root.findViewById<TextView>(R.id.text_a1c_control_avg)
         val textControlBad = root.findViewById<TextView>(R.id.text_a1c_control_bad)
+        val textNoData = root.findViewById<TextView>(R.id.text_no_data)
+        val textError = root.findViewById<TextView>(R.id.text_error)
 
-        model.getStatsObserver().observe(viewLifecycleOwner, { stats ->
-            textMin.text = stats.minFormatted
-            textMax.text = stats.maxFormatted
-            textAvg.text = stats.avgFormatted
-            textA1C.text = stats.a1cFormatted
+        viewModel.statsLiveData().observe(viewLifecycleOwner, { stats ->
+            if (stats != null) {
+                textMin.text = stats.minFormatted
+                textMax.text = stats.maxFormatted
+                textAvg.text = stats.avgFormatted
+                textA1C.text = stats.a1cFormatted
+                content.show()
+                textNoData.hide()
+            } else {
+                content.hide()
+                textNoData.show()
+            }
         })
-        model.getDiabetesControlObserver().observe(viewLifecycleOwner, {
+
+        viewModel.showErrorLiveData().observe(viewLifecycleOwner, {
+            setViewVisible(it, textError)
+        })
+
+        viewModel.diabetesControlLiveData().observe(viewLifecycleOwner, {
             hideView(textControlGood, textControlAvg, textControlBad)
             when(it) {
                 0 -> textControlGood.show()
@@ -43,7 +56,7 @@ class StatisticsFragment: Fragment() {
         })
 
         spinnerPeriod.onItemSelectedListener = setItemSelectedListener {
-            model.setPeriod(it)
+            viewModel.setPeriod(it)
         }
         return root
     }